+++ Shooterman/Server/Systems/CollisionSystem/Collision.cpp
@@ -66,11 +66,11 @@ namespace Collision
       sf::Uint8* mask2 = Bitmasks.GetMask(Object2.getTexture());
 
       // Loop through our pixels
-      for (int i = Intersection.left; i < Intersection.left + Intersection.width; i++) {
-        for (int j = Intersection.top; j < Intersection.top + Intersection.height; j++) {
+      for (int i = (int)(Intersection.left); i < Intersection.left + Intersection.width; i++) {
+        for (int j = (int)(Intersection.top); j < Intersection.top + Intersection.height; j++) {
 
-          sf::Vector2f o1v = Object1.getInverseTransform().transformPoint(i, j);
-          sf::Vector2f o2v = Object2.getInverseTransform().transformPoint(i, j);
+          sf::Vector2f o1v = Object1.getInverseTransform().transformPoint((float)i, (float)j);
+          sf::Vector2f o2v = Object2.getInverseTransform().transformPoint((float)i, (float)j);
 
           // Make sure pixels fall within the sprite's subrect
           if (o1v.x > 0 && o1v.y > 0 && o2v.x > 0 && o2v.y > 0 &&
@@ -132,9 +132,9 @@ namespace Collision
       sf::Transform trans = Object.getTransform();
       sf::IntRect local = Object.getTextureRect();
       Points[0] = trans.transformPoint(0.f, 0.f);
-      Points[1] = trans.transformPoint(local.width, 0.f);
-      Points[2] = trans.transformPoint(local.width, local.height);
-      Points[3] = trans.transformPoint(0.f, local.height);
+      Points[1] = trans.transformPoint((float)(local.width), 0.f);
+      Points[2] = trans.transformPoint((float)(local.width), (float)(local.height));
+      Points[3] = trans.transformPoint(0.f, (float)(local.height));
     }
 
     sf::Vector2f Points[4];
